requirements.txt
pandas
numpy
plotly.express
seaborn
google.oauth2 
pandas_gbq
dotenv
os